+++ src/views/components/Card.vue
@@ -46,6 +46,7 @@ export default {
   padding: 0;
   border-radius: 3px;
   margin-bottom: 30px;
+  box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
 }
 
 .c-card__header {