+++ src/views/components/ColumnDetails/AddressColumn.vue
@@ -1,0 +1,22 @@
+<template>
+  <div>
+    <div>
+      {{ $options.formatAddress((data.student && data.student.address) || {}) }}
+    </div>
+  </div>
+</template>
+
+<script>
+  import { formatAddress } from '../../../helpers/utils';
+  export default {
+    formatAddress,
+    props: {
+      data: {
+        type: [Object]
+      }
+    }
+  }
+</script>
+<style lang="scss" scoped>
+  @import "../../../assets/scss/shared.scss";
+</style>