+++ src/views/components/ColumnDetails/index.js
@@ -1,3 +1,4 @@
 export { default as StudentColumn } from './StudentColumn';
 export { default as EducationColumn } from './EducationColumn';
 export { default as ContactColumn } from './ContactColumn';
+export { default as AddressColumn } from './AddressColumn';